requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests~=2.32.5
APScheduler==3.10.4
Jinja2==3.1.4
pytz==2024.1
matplotlib==3.9.0
pandas==2.2.2
openpyxl==3.1.3
python-telegram-bot==21.4
feedparser==6.0.11
